Shahar: merging & refactoring to the right directory
