+++ src/main/java/frc/robot/Robot.java
@@ -6,8 +6,8 @@ package frc.robot;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
-import frc.robot.poseestimator.photonvision.PhotonVisionConstants;
-import frc.robot.poseestimator.photonvision.photonvisionfilters.PhotonVisionAprilTagFilters;
+import frc.robot.vision.photonvision.PhotonVisionConstants;
+import frc.robot.vision.photonvision.photonvisionfilters.PhotonVisionAprilTagFilters;
 
 /**
  * This class is where the bulk of the robot should be declared. Since Command-based is a "declarative" paradigm, very little
