+++ src/main/java/frc/robot/vision/photonvision/CameraConfiguration.java
@@ -1,4 +1,4 @@
-package frc.robot.poseestimator.photonvision;
+package frc.robot.vision.photonvision;
 
 import edu.wpi.first.math.geometry.Transform3d;
